+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.3.61
+
+- Hotfix
+
+## 3.3.60
+
+- Added `readyToOperate` and `unloadingOperation` values in `AtsPurchaseOrderStatus` enum
+- Added `unloadingOperationAt` and `readyToOperateAt` in `AtsOperationStatuses` model
+
 ## 3.3.59
 
 - Added `billingPlan` to `User` model
@@ -14,6 +23,7 @@
 - Added `aiEnabled` to `BillingPlan` and their `BillingPlanInput` models
 
 ## 3.3.56
+
 - Added `operationId` in `AtsEntry`
 
 ## 3.3.55
@@ -21,9 +31,11 @@
 - Added lock, suspension and paymentWarning fields to `User` model
 
 ## 3.3.54
+
 - updated `TriggerInput` model to mathch the new `Trigger` model
 
 ## 3.3.53
+
 - added in `isBluetooth` in `AtsEntryUnloadInput`
 
 ## 3.3.52